+++ client/src/components/posts/postList.jsx
@@ -1,15 +1,15 @@
-import React, { Component } from "react";
-import { connect } from "react-redux";
-import axios from "axios";
-import { bindActionCreators } from "redux";
-import { updatePosts } from "../actions/index.jsx";
-import PostDetails from "./postDetails.jsx";
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import axios from 'axios';
+import { bindActionCreators } from 'redux';
+import { updatePosts } from './../../actions/index.jsx';
+import PostDetails from './postDetails.jsx';
 
 class PostList extends Component {
   componentDidMount() {
     axios
-      .get("/posts")
-      .then(res => {
+      .get('/posts')
+      .then((res) => {
         this.props.updatePosts(res.data);
       })
       .catch((err) => {
@@ -24,11 +24,7 @@ class PostList extends Component {
 
     return (
       <ul>
-        {this.props.posts
-          .map((post, index) => {
-            return <PostDetails post={post} key={index} />;
-          })
-          .reverse()}
+        {this.props.posts.map((post, index) => <PostDetails post={post} key={index} />).reverse()}
       </ul>
     );
   }
